Fix implicit long-to-int narrowing in GwcFixture's fake SeekableFile
CodeQL flagged pos += what (pos is int, what is long) as an implicit
narrowing conversion in skip(). seek() already cast explicitly for the
same reason; skip() should too.

+++ wherigo/src/test/java/cgeo/geocaching/wherigo/openwig/formats/GwcFixture.java
@@ -151,7 +151,7 @@ final class GwcFixture {
         }
 
         public long skip(final long what) {
-            pos += what;
+            pos += (int) what;
             return what;
         }
